Tic-Tac-Toe with Created Grid
Creating a Tic-Tac-Toe display board grid, then having one of the two players chosen at random to go first. Players continue to take turns choosing the space they want to place their mark. The code then checks to make sure the space is available and checks if the player won with that position placement. If there is no winner then the next players turn until there is a winner or there is no more spaces to play.